refactor: `ProfileModal` 경로 수정
`Modal`에 대한 의존성과 비슷한 컴포넌트끼리 묶어두기 위해
