+++ src/components/common/Modal/ProfileModal/index.ts
@@ -1,0 +1,3 @@
+export { default as ProfileModal } from './ProfileModal';
+export { default as ProfileButton } from './ProfileButton';
+export { default as ProfileContent } from './ProfileContent';
